+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/FiveFretGuitar/GuitarHeroGuitar/PS3GuitarHeroGuitar.cs
@@ -100,10 +100,18 @@ namespace PlasticBand.Devices
         /// </summary>
         internal new static void Initialize()
         {
+            // PS3 guitars
             InputSystem.RegisterLayout<PS3GuitarHeroGuitar>(matches: new InputDeviceMatcher()
                 .WithInterface("HID")
                 .WithCapability("vendorId", 0x12BA)
                 .WithCapability("productId", 0x0100)
+            );
+
+            // World Tour PC guitar
+            InputSystem.RegisterLayout<PS3GuitarHeroGuitar>(matches: new InputDeviceMatcher()
+                .WithInterface("HID")
+                .WithCapability("vendorId", 0x1430)
+                .WithCapability("productId", 0x474C)
             );
         }